+++ demo/stores/app.js
@@ -137,6 +137,7 @@ let data = ImmutableHelper.parseJSON({
   },
   pod: {
     border: true,
+    editContentFullWidth: true,
     padding: "medium",
     as: "primary",
     alignTitle: 'left'